+++ Packages/com.github.kurotu.vrc-quest-tools/Runtime/Components/ConvertedAvatar.cs
@@ -1,0 +1,14 @@
+﻿using UnityEngine;
+using VRC.SDKBase;
+
+namespace KRT.VRCQuestTools.Components
+{
+    /// <summary>
+    /// Component to mark converted avatar.
+    /// </summary>
+    [AddComponentMenu("VRCQuestTools/VQT Converted Avatar")]
+    [DisallowMultipleComponent]
+    public class ConvertedAvatar : MonoBehaviour, IEditorOnly
+    {
+    }
+}